Refactor flock objects to inherit from closure_collector core objects
Added mapping and sequence support natively to `closure_collector.core` with `ClosureMapping` and `ClosureList` (which extend `ClosurePromiseMapping`).
Updated `flock.core` objects (`FlockDict`, `FlockList`, `PromiseFlock`, `FlockAggregator`) to inherit from these new base classes while preserving their exact backwards-compatible interface (`FlockException` etc.).
Migrated core logic and complexity previously isolated to `flock` into the `closure_collector` namespace cleanly.
